move test_reset_limits into test_viewer.py
